Factor out a SearchRequester class.
